feat(observers): Implement observer execution logging to tb_observer_log
Implemented observer execution logging that writes results to the
tb_observer_log table after processing each event.

Runtime Changes:
- Modified load_observers() to return entity_type index for lookup
- After processing events, write logs with status, duration, errors
- Fixed cleanup order to prevent foreign key violations

Test Results:
8 out of 14 tests now passing (was 7 before)

New passing tests:
- test_debug_event_processing
- test_high_throughput_processing
- test_runtime_continues_after_errors

Remaining failures are checkpoint-related, not logging issues.

+++ crates/fraiseql-server/tests/observer_test_helpers.rs
@@ -522,7 +522,7 @@ pub async fn check_checkpoint_exists(
     pool: &PgPool,
     listener_id: &str,
 ) -> Result<bool, sqlx::Error> {
-    let row: Option<(i32,)> = sqlx::query_as(
+    let row: Option<(i64,)> = sqlx::query_as(
         "SELECT COUNT(*) FROM observer_checkpoints WHERE listener_id = $1",
     )
     .bind(listener_id)
